requirements.txt
streamlit
groq
google-generativeai
duckduckgo-search
python-dotenv